wire the delagates and usecases

+++ app/src/main/java/de/rwth_aachen/phyphox/features/settings/presentation/viewmodel/delegates/graphsize/GraphSizeDelegate.kt
@@ -8,5 +8,5 @@ import kotlinx.coroutines.flow.Flow
 interface GraphSizeDelegate {
     val graphSizeFlow: Flow<UiResourceState<SeekBarConfig>>
     fun start(scope: CoroutineScope)
-    suspend fun updateGraphSize(size: Float)
+    suspend fun updateGraphSize(size: Int)
 }
